+++ CHANGELOG.md
@@ -147,4 +147,7 @@ Changes adapted from [homebridge-neato](https://github.com/naofireblade/homebrid
 * Changed config parameter **hidden**. Renamed to **services**, now takes list of services that should be _visible_. Default are all available services.
 * Fixed robots no longer disappear or change the room after connection issues with the Vorwerk API
 * Fixed plugin no longer crashes if non smart robot is assigned in Vorwerk account
-* Fixed options for eco, nogo lines, extra care are now saved in homebridge and will no longer be overridden by Vorwerk API
+* Fixed options for eco, nogo lines, extra care, spot repeat, spot size are now saved in homebridge and will no longer be overridden by Vorwerk API
+
+## TODO until 1.0.0 release
+ * Room cleaning